refactor: json object cls struct docstring example
Moved usage example of JsonObjectClassStruct from inline to a module (literalinclude), and added it to docstring tests.
